♻️ refactor(migration): スクリプトを簡素化し export/transform/import の3コマンドに統一
複雑化していた移行スクリプトを整理し、コア機能のみに削減:

- 16ファイル → 7ファイル (56%削減)
- ~5,000行 → ~1,375行 (72%削減)
- 8コマンド → 3コマンド (export/transform/import)

削除した機能:
- discover: 環境自動検出
- backup: DynamoDB オンデマンドバックアップ
- verify: データ検証
- report: レポート生成
- full/init: 複合コマンド

残した機能:
- export: Bot テーブルを JSON エクスポート
- transform: Bot → Assistant 形式変換
- import: DynamoDB 投入 + S3 コピー

マルチテナント・クロスアカウント対応は維持
