+++ packages/cdk/scripts/migration/transform/botToAssistant.ts
@@ -1,107 +1,16 @@
-#!/usr/bin/env node
-
 /**
- * Bot -> Assistant Transform Script
+ * Bot -> Assistant Transform
  * v0.5.3 の BotTableV3 データを develop の Assistant 形式に変換
  */
 
-import * as fs from 'fs';
-import * as path from 'path';
 import { randomUUID } from 'crypto';
-
-// ============================================================================
-// Types
-// ============================================================================
-
-/**
- * v0.5.3 BotTableV3 の Bot アイテム
- */
-interface V053Bot {
-  PK: string; // UserId
-  SK: string; // BOT#{bot_id}
-  ItemType?: string;
-  BotId: string;
-  Title: string;
-  Description?: string;
-  Instruction?: string;
-  CreateTime?: number;
-  LastUsedTime?: number;
-  SharedScope?: 'private' | 'partial' | 'all';
-  SharedStatus?: string;
-  SyncStatus?: string;
-  SyncStatusReason?: string;
-  Knowledge?: {
-    filenames?: string[];
-    source_urls?: string[];
-    sitemap_urls?: string[];
-    s3_urls?: string[];
-  };
-  ConversationQuickStarters?: Array<{ title: string; example?: string }>;
-  BedrockKnowledgeBase?: Record<string, unknown>;
-  // その他のフィールドは必要に応じて追加
-}
-
-/**
- * develop の Assistant 形式
- */
-interface AssistantItem {
-  id: string; // userId - partition key
-  createdDate: string; // sort key
-  assistantId: string;
-  userId: string;
-  tenantId: string;
-  name: string;
-  description: string;
-  instruction: string;
-  modelId: string;
-  ragEnabled: boolean;
-  visibility: 'private' | 'public';
-  syncStatus: 'QUEUED' | 'SYNCING' | 'SUCCEEDED' | 'FAILED' | 'PARTIAL';
-  syncStatusReason: string;
-  knowledgeSources: KnowledgeSource[];
-  firstQuestions?: string[];
-  s3Urls?: string[];
-  updatedDate: string;
-}
-
-interface KnowledgeSource {
-  id: string;
-  type: 'file' | 'web' | 'url';
-  name: string;
-  displayName?: string;
-  storageKey?: string;
-  sourceUrl?: string;
-  status: 'QUEUED' | 'SYNCING' | 'SUCCEEDED' | 'FAILED';
-}
-
-/**
- * S3 コピーマッピング
- */
-interface S3CopyMapping {
-  sourceKey: string;
-  targetKey: string;
-  fileName: string;
-  fileId: string;
-  botId: string;
-  userId: string;
-}
-
-/**
- * 変換結果
- */
-interface TransformResult {
-  assistants: AssistantItem[];
-  s3Mappings: S3CopyMapping[];
-  botIdToAssistantId: Record<string, string>;
-  statistics: {
-    totalBots: number;
-    transformedAssistants: number;
-    totalFiles: number;
-    totalUrls: number;
-    skipped: number;
-    errors: string[];
-  };
-}
+import {
+  V053Bot,
+  AssistantItem,
+  KnowledgeSource,
+  S3CopyMapping,
+  TransformResult,
+} from '../types';
 
 // ============================================================================
 // Transform Functions
@@ -298,146 +207,4 @@ export function transformBots(
       errors,
     },
   };
-}
-
-// ============================================================================
-// CLI
-// ============================================================================
-
-interface CliOptions {
-  input: string;
-  output: string;
-  tenantId: string;
-  modelId?: string;
-}
-
-function parseArgs(): CliOptions {
-  const args = process.argv.slice(2);
-  const options: Partial<CliOptions> = {};
-
-  for (let i = 0; i < args.length; i++) {
-    switch (args[i]) {
-      case '-i':
-      case '--input':
-        options.input = args[++i];
-        break;
-      case '-o':
-      case '--output':
-        options.output = args[++i];
-        break;
-      case '-t':
-      case '--tenant-id':
-        options.tenantId = args[++i];
-        break;
-      case '-m':
-      case '--model-id':
-        options.modelId = args[++i];
-        break;
-      case '-h':
-      case '--help':
-        printHelp();
-        process.exit(0);
-    }
-  }
-
-  if (!options.input || !options.output || !options.tenantId) {
-    console.error('Error: Missing required arguments');
-    printHelp();
-    process.exit(1);
-  }
-
-  return options as CliOptions;
-}
-
-function printHelp(): void {
-  console.log(`
-Bot to Assistant Transform Script
-
-Usage:
-  npx ts-node transform/botToAssistant.ts [options]
-
-Options:
-  -i, --input <path>      Input JSON file (DynamoDB scan result)
-  -o, --output <path>     Output directory
-  -t, --tenant-id <id>    Tenant ID
-  -m, --model-id <id>     Default model ID (optional)
-  -h, --help              Show help
-
-Example:
-  npx ts-node transform/botToAssistant.ts \\
-    -i ./backup/bots.json \\
-    -o ./output \\
-    -t my-tenant
-`);
-}
-
-async function main(): Promise<void> {
-  const options = parseArgs();
-
-  console.log('=== Bot to Assistant Transform ===');
-  console.log(`Input: ${options.input}`);
-  console.log(`Output: ${options.output}`);
-  console.log(`Tenant ID: ${options.tenantId}`);
-
-  // 入力ファイルを読み込み
-  const inputData = JSON.parse(fs.readFileSync(options.input, 'utf-8'));
-
-  // DynamoDB scan 結果の場合、Items を抽出
-  const bots: V053Bot[] = inputData.Items || inputData;
-
-  console.log(`\nFound ${bots.length} items in input file`);
-
-  // 変換実行
-  const result = transformBots(bots, options.tenantId, options.modelId);
-
-  // 出力ディレクトリを作成
-  if (!fs.existsSync(options.output)) {
-    fs.mkdirSync(options.output, { recursive: true });
-  }
-
-  // 結果を保存
-  const assistantsPath = path.join(options.output, 'assistants.json');
-  const s3MappingsPath = path.join(options.output, 's3-mappings.json');
-  const idMappingPath = path.join(options.output, 'id-mapping.json');
-  const statsPath = path.join(options.output, 'transform-stats.json');
-
-  fs.writeFileSync(assistantsPath, JSON.stringify(result.assistants, null, 2));
-  fs.writeFileSync(s3MappingsPath, JSON.stringify(result.s3Mappings, null, 2));
-  fs.writeFileSync(
-    idMappingPath,
-    JSON.stringify(result.botIdToAssistantId, null, 2)
-  );
-  fs.writeFileSync(
-    statsPath,
-    JSON.stringify(result.statistics, null, 2)
-  );
-
-  // 統計を表示
-  console.log('\n=== Transform Statistics ===');
-  console.log(`Total Bots: ${result.statistics.totalBots}`);
-  console.log(`Transformed Assistants: ${result.statistics.transformedAssistants}`);
-  console.log(`Total Files: ${result.statistics.totalFiles}`);
-  console.log(`Total URLs: ${result.statistics.totalUrls}`);
-  console.log(`Skipped: ${result.statistics.skipped}`);
-
-  if (result.statistics.errors.length > 0) {
-    console.log(`\nErrors (${result.statistics.errors.length}):`);
-    result.statistics.errors.forEach((e) => console.log(`  - ${e}`));
-  }
-
-  console.log('\n=== Output Files ===');
-  console.log(`Assistants: ${assistantsPath}`);
-  console.log(`S3 Mappings: ${s3MappingsPath}`);
-  console.log(`ID Mapping: ${idMappingPath}`);
-  console.log(`Statistics: ${statsPath}`);
-
-  console.log('\nTransform completed successfully!');
-}
-
-// 直接実行された場合のみ main() を呼び出す
-if (require.main === module) {
-  main().catch((error) => {
-    console.error('Transform failed:', error);
-    process.exit(1);
-  });
 }
